pull SubmissionDetail into its own file
